Create  Strong Primes
You are given Q queries and, in each query, there are two numbers L and R. You have to calculate the number of strong primes present in the range L and R inclusive.
Note
A strong prime is a prime number that is greater than the arithmetic mean of the nearest prime above and below. Algebraically, a prime P said to be strong if
2Pn > Pn-1 + Pn+1,
where n is their index in the ordered set of prime numbers, where Pi denotes the ith. prime.
I
•	The first line of the input contains an integer Q denoting the number of Queries.
•	Then Q lines follow each containing two numbers L and R.
Output format
•	For each query, print the number of Strong primes present in the range L to R inclusively. The answer to each test case should come in a new line.
Constraint:
•	1<=Q<=105
•	1<=L<=R<=106
Sample Input:
	3 
	10 20
	20 30
	30 50
Sample Output:
	2
	1
	2
Explanation:
•	For the range 10 to 20, there are 2 strong primes (11, 17).
•	For the range 20 to 30, there is 1 strong prime (29).
•	For the range 30 to 50, there are 2 strong primes (37, 41).