Create HackWithInfy Weird Elevators
A multi-story building has several floors. Each floor has an elevator. You are given two integers X and Y. The elevators in the building have a system with the following rules:
•	From any floor number A, you can only move to the floors numbered A/P, where P is a prime factor of A that is smaller than a given number M
•	The elevator takes exactly 1 second to move to another floor
•	Only one elevator can move at a time.
Write a program to find the minimum time required to move to a common floor.
Input format
•	First line: T (number of test cases)
•	Next T lines: Three space-separated integers X, Y and M
Output format
For each test case, print 2 space-separated integers in a single line, i.e., the time taken to meet and the floor number at which you meet. In case it is not possible to meet, print -1
Constraints
•	1 <= T <= 100
•	1 <= X,Y <= 106
•	1 <= M <= 106
Sample Input
	3
	20 16 10
	100 120 10
	160 180 10
Sample Output
	3 4
	3 20
	5 20
Explanation
Case 1:
•	After 1 second: You move from 20 to 4 (dividing by 5)
•	After 2 seconds: Your friend moves from 16 to 8 (dividing by 2)
•	After 3 seconds: Your friend moves from 8 to 4 (dividing by 2)